requirements.txt
requests
flask
argparse
pyaml